pyproject.toml
[tool.poetry]
name = "productlab_test_task"
version = "0.1.0"
description = ""
authors = ["PashkovichKaterina <pashkovichkate@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
aiohttp = "^3.8.1"
openpyxl = "^3.0.10"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
